Reformat YANG files using pyang -f yang.
This change reformats the YANG models in the openhltest repository
programmatically using pyang -f yang. This ensures that the indentation
style is consistent throughout. There are some whitespace lines removed
which may or may not be desirable, but currently there is a strange mix
of indendation through the project which hurts readability.